Wrote in each set for columns + tag for log output in console.
Unable it seems to split the columns due to the data not have in place the
spilt comma, however tried regex which gave back better response to the file.
Can't use different columns to display data in app, have tried different methods
to do this, however best results is to allow all data to show, which is not needed
for test. Remote csv file working and have left in place for testing the raw folder
 ith copy of csv file.